src/main/java/com/codility/lesson/arrays/OddOccurrencesInArray.java
package com.codility.lesson.arrays;

import java.util.HashMap;
import java.util.Map;

public class OddOccurrencesInArray {
    public int solution(int[] A) {
        Map<Integer, Integer> countMap = new HashMap<>();
        for (int i = 0; i < A.length; i++) {
            if (!countMap.containsKey(A[i])) {
                countMap.put(A[i], 0);
            }

            countMap.put(A[i], countMap.get(A[i]) + 1);
        }

        for (Map.Entry<Integer, Integer> entry : countMap.entrySet()) {
            if (entry.getValue() % 2 == 1) {
                return entry.getKey();
            }
        }

        return -1;
    }
}
